+++ src/services/auditLog.service.js
@@ -17,4 +17,36 @@ export default class AuditLogService extends Service{
               error(err);
           });
     }
+
+    findByAuditIdFiltered(id, sharecode, page, size, asc, levels, then, error){
+        
+        var levelsData = [];
+        levels.forEach(element => levelsData.push('&levels='+element));
+        levelsData = levelsData.join("");
+
+        return this.axios({
+            url: this.controllerName + '/by-audit-filtered/' + id + '/' + sharecode + '?page=' + page + '&size=' + size + '&asc=' + asc
+            + levelsData,
+            method: 'get',
+          })
+          .then(resp => {
+            then(resp.data)
+          })
+          .catch(err => {
+              error(err);
+          });
+    }
+
+    getLogLevels(then, error){
+        return this.axios({
+            url: this.controllerName + '/levels',
+            method: 'get',
+          })
+          .then(resp => {
+            then(resp.data)
+          })
+          .catch(err => {
+              error(err);
+          });
+    }
 }