+++ src/views/AuditDetail/LogsList.vue
@@ -1,6 +1,5 @@
 <template>
     <div>
-        <h2>{{ $t('auditDetail.logs') }} ({{totalElements}})</h2>
         <div class="list-audits" role="list">
             <p class="list-audits__item" role="listitem" v-for="(auditLog, i) in auditLogs" :key="auditLog.id">
                 <span class="list-audits__number">{{minElementIndex + i}}.</span> <b>{{auditLog.level}}</b> - [{{$moment(auditLog.date).format('DD-MM-YYYY HH:mm')}}] {{auditLog.message}}
@@ -11,7 +10,7 @@
 
 <script>
 export default {
-	name: "logsList",
+    name: "logsList",
     props: ['audit', 'auditLogs', 'totalElements', 'currentPage', 'elementByPage'],
     computed: {
 	    minElementIndex(){