+++ CHANGELOG.md
@@ -29,3 +29,4 @@ Improvements :
 - Improved RGAA text display
 - Improved result diplay in synthesis table pop-ins
 - Admin and SuperAdmin can unlock/lock users
+- Audit logs can be filtered by level and sort by date